refactor: reorg ShipDb.lua, PilotDb.lua, and UpgradeDb.lua to Spawner/
